requirements.txt
certifi==2019.9.11
chardet==3.0.4
codecov==2.0.15
coverage==4.5.4
Django==2.2.5
django-debug-toolbar==2.0
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.8
idna==2.8
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
